+++ src/algorithms/constructors/mod.rs
@@ -13,6 +13,7 @@ pub mod hamming;
 pub mod hypercube;
 pub mod kary_tree;
 pub mod kautz;
+pub mod lcf;
 pub mod linegraph;
 pub mod prufer;
 pub mod regular_tree;